refactor(domain): retire facade.py — single boundary is now terok.lib.api
Drop the transitional re-export shim that previously fronted the domain
layer.  Its contents were carved into ``terok.lib.domain.{auth,ssh}``
and ``terok.lib.domain.project`` factory functions in the preceding
commits; api.py now imports each directly and the CLI consumes api.

Mechanical cleanup:

- Delete ``src/terok/lib/domain/facade.py``.
- Drop the ``terok.lib.domain.facade`` module + interface blocks from
  ``tach.toml`` and the matching entries from ``.importlinter``'s
  sandbox-boundary and executor-boundary allowed_importers (the
  consolidation the file's preamble was hoping for).
- Rename ``tests/unit/lib/domain/test_facade.py`` to ``test_domain_api.py``
  so the file name matches what it actually tests now.
- Re-point a handful of test imports (``test_cli_workflows.py``,
  ``test_project_archive.py``, ``test_project_delete.py``) from facade
  to ``terok.lib.api``.

After this commit terok has one public boundary (``terok.lib.api``)
shared by CLI and TUI, no leaky "facade" re-export module, and the
``allowed_importers`` lists are a step closer to the eventual
``terok.lib.integrations.*``-only floor planned for Phase 3.
