+++ tests/unit/lib/test_project_delete.py
@@ -11,10 +11,10 @@ from types import SimpleNamespace
 
 import pytest
 
+from terok.lib.api import delete_project
 from terok.lib.core.config import build_dir as cfg_build_dir
 from terok.lib.core.paths import core_state_dir as cfg_state_dir
 from terok.lib.core.projects import load_project
-from terok.lib.domain.facade import delete_project
 from tests.test_utils import project_env, write_project
 
 EnvSetup = Callable[[SimpleNamespace, str], Path]
